Preserve trailing comma in EXTINF directives
The parser should preserve trailing commas in EXTINF directives.
According to 4.3.2.1 of the HLS spec, the comma is required as the
EXTINF format is specified as:

    #EXTINF:<duration>,[<title>]

To preserve this, the EXTINF parameters are always returned as a
string. They are not attempted to be parsed as an attribute list.
This allows source playlists with trailing commas to preserve them
during parsing, and vice-versa.